refactor: 내가 쓴 자유게시판 글 관련 기능 BoardController로 이동
